feature: LadderLine의 두 점과 현재 위치를 Point로 추상화
Point 리팩토링
 - primitive type인 index를 Index 클래스로

+++ src/main/java/nextstep/ladder/domain/LadderLine.java
@@ -11,7 +11,7 @@ public class LadderLine {
     }
 
     public int move(int position) {
-        return points.get(position).move();
+        return points.get(position).move().toInt();
     }
 
     public static LadderLine init(int sizeOfPerson, LadderPointGenerateStrategy strategy) {
